refactor: move logic to member section
